+++ src/create/mod.rs
@@ -1,11 +1,6 @@
-pub mod basic;
-pub mod math;
-pub mod book;
+pub mod routes;
 pub mod config;
-pub mod theatre;
 
-pub use math::*;
-pub use book::*;
+
+pub use routes::*;
 pub use config::*;
-pub use basic::*;
-pub use theatre::*;