+++ RELEASES.md
@@ -3,3 +3,8 @@
 - Fixed list command 
 - Now expects uppercased template name
 - Added init command
+
+# Version 0.3.0
+- Added document options in config.toml
+- Reduced code repititon with routes.rs
+- All templates with a new main.tex & structure.tex standard